+++ src/app/app.js
@@ -21,15 +21,22 @@ function renderComponent(react, id) {
 }
 
 function Warning(props) {
-  const { title, preformatted, unformatted } = props;
+  const {
+    title,
+    preformatted,
+    unformatted,
+    theme,
+  } = props;
   return (
-    <div className="container-fluid">
-      <div className="row">
-        <div className="col-12">
-          <div className={LIGHT_CARD}>
-            <h1>{title}</h1>
-            <pre>{preformatted}</pre>
-            <div>{unformatted}</div>
+    <div className={`vitessce-container vitessce-theme-${theme}`}>
+      <div className="warning-layout container-fluid">
+        <div className="row">
+          <div className="col-12">
+            <div className={LIGHT_CARD}>
+              <h1>{title}</h1>
+              <pre>{preformatted}</pre>
+              <div>{unformatted}</div>
+            </div>
           </div>
         </div>
       </div>
@@ -47,13 +54,14 @@ function preformattedDetails(response) {
     url: ${response.url}`; // TODO: headers
 }
 
-export function validateAndRender(config, id, rowHeight) {
+export function validateAndRender(config, id, rowHeight, theme) {
   if (!config) {
     // If the config value is undefined, show a warning message
     renderComponent(
       <Warning
         title="No such dataset"
         unformatted="The dataset configuration could not be found."
+        theme={theme}
       />, id,
     );
     return;
@@ -71,6 +79,7 @@ export function validateAndRender(config, id, rowHeight) {
       <Warning
         title="Config validation failed"
         preformatted={failureReason}
+        theme={theme}
       />, id,
     );
     return;
@@ -80,29 +89,32 @@ export function validateAndRender(config, id, rowHeight) {
       config={config}
       getComponent={getComponent}
       rowHeight={rowHeight}
+      theme={theme}
     />, id,
   );
 }
 
-function renderResponse(response, id) {
+function renderResponse(response, id, theme) {
   if (!response.ok) {
     renderComponent(
       <Warning
         title="Fetch response not OK"
         preformatted={preformattedDetails(response)}
+        theme={theme}
       />, id,
     );
   } else {
     response.text().then((text) => {
       try {
         const config = JSON.parse(text);
-        validateAndRender(config, id);
+        validateAndRender(config, id, undefined, theme);
       } catch (e) {
         renderComponent(
           <Warning
             title="Error parsing JSON"
             preformatted={preformattedDetails(response)}
             unformatted={`${e.message}: ${text}`}
+            theme={theme}
           />, id,
         );
       }
@@ -110,26 +122,37 @@ function renderResponse(response, id) {
   }
 }
 
+/**
+ * Use the theme provided if it is valid, otherwise fall back to the 'dark' theme.
+ * @param {string} theme A potentially invalid theme name.
+ * @returns {string} A valid theme name.
+ */
+function validateTheme(theme) {
+  return (['light', 'dark'].includes(theme) ? theme : 'dark');
+}
+
 export function renderApp(id, rowHeight = null) {
   const urlParams = new URLSearchParams(window.location.search);
   const datasetId = urlParams.get('dataset');
   const datasetUrl = urlParams.get('url');
   const showAll = urlParams.get('show') === 'all';
+  const theme = validateTheme(urlParams.get('theme'));
 
   if (datasetId) {
     const config = getConfig(datasetId);
-    validateAndRender(config, id, rowHeight);
+    validateAndRender(config, id, rowHeight, theme);
   } else if (datasetUrl) {
     fetch(datasetUrl)
-      .then(response => renderResponse(response, id))
+      .then(response => renderResponse(response, id, theme))
       .catch(error => renderComponent(
         <Warning
           title="Error fetching"
           unformatted={error.message}
+          theme={theme}
         />, id,
       ));
   } else {
     const configs = listConfigs(showAll);
-    renderComponent(<Welcome configs={configs} />, id);
+    renderComponent(<Welcome configs={configs} theme={theme} />, id);
   }
 }